refactor(tui): extract plugin source watcher
